backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.17,<4.0
django-cors-headers>=4.9,<5.0
djangorestframework-simplejwt>=5.5,<6.0
psycopg[binary]>=3.3,<4.0
python-dotenv>=1.2,<2.0

